Changing the files as per my new minimalist macros.

+++ Template Library/CHT.cpp
@@ -1,25 +1,25 @@
 // still need to test
 struct CHT {
     struct Line {
-        ll m, c;
+        int m, c;
 
-        Line(ll a, ll b) {
+        Line(int a, int b) {
             m = a;
             c = b;
         }
 
-        ll isect(Line &other) {
+        int isect(Line &other) {
             return (c - other.c + other.m - m - 1) / (other.m - m);
         }
 
-        ll val(ll x) {
+        int val(int x) {
             return m*x + c;
         }
     };
 
-    deque<pair<Line, ll>> dq;
+    deque<pair<Line, int>> dq;
 
-    void insert(ll m, ll c) {
+    void insert(int m, int c) {
         Line nx = Line(m, c);
         if(dq.empty()) {
             dq.eb(nx, 0);
@@ -28,7 +28,7 @@ struct CHT {
 
         while(!dq.empty()) {
             Line pr = dq.back().fr;
-            ll x = dq.back().sc;
+            int x = dq.back().sc;
             if(nx.m == pr.m) {
                 if(nx.c > pr.c) {
                     return;
@@ -47,8 +47,8 @@ struct CHT {
         dq.eb(nx, nx.isect(dq.back().fr));
     }
 
-    ll query(ll x) {
-        ll m = dq.size();
+    int query(int x) {
+        int m = dq.size();
         while(m > 1 && dq[1].sc <= x) {
             dq.pop_front();
             --m;
@@ -56,17 +56,16 @@ struct CHT {
         return dq.front().fr.val(x);
     }
 
-    ll query2(ll x) {
-        ll lo = 0;
-        ll hi = dq.size();
+    int query2(int x) {
+        int lo = 0;
+        int hi = dq.size();
         while(lo + 1 != hi) {
-            ll mid = (lo + hi) / 2;
+            int mid = (lo + hi) / 2;
             if(dq[mid].sc <= x)
                 lo = mid;
             else
                 hi = mid;
         }
-
         return dq[lo].fr.val(x);
     }
 };
